requirements.txt
pandas==2.1.0
pywhatkit==5.4
openpyxl==3.1.2
python-docx==0.8.11
tk==0.1.0
pip